requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
click==8.1.7
crispy-bootstrap4==2024.1
Django==5.0.4
django-bootstrap4==24.3
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-haystack==3.2.1
django-js-asset==2.2.0
djangorestframework==3.15.1
joblib==1.4.0
nltk==3.8.1
pillow==10.3.0
psycopg==3.1.18
regex==2024.4.16
soupsieve==2.5
sqlparse==0.5.0
tqdm==4.66.2
typing_extensions==4.11.0
Whoosh==2.7.4
